hls: fixed some bugs, live stream caching starts only after first hls request
<==== Mercurial data ====>
Branch: hls

+++ mediaserver/src/streaming/hls/hls_archive_playlist_manager.cpp
@@ -77,7 +77,7 @@ namespace nx_hls
         return true;
     }
 
-    unsigned int ArchivePlaylistManager::generateChunkList(
+    size_t ArchivePlaylistManager::generateChunkList(
         std::vector<AbstractPlaylistManager::ChunkData>* const chunkList,
         bool* const endOfStreamReached ) const
     {
