+++ mediaserver/src/streaming/hls/hls_playlist_manager.h
@@ -38,7 +38,7 @@ namespace nx_hls
             \param endOfStreamReached Can be NULL
             \return Number of chunks generated
         */
-        virtual unsigned int generateChunkList(
+        virtual size_t generateChunkList(
             std::vector<ChunkData>* const chunkList,
             bool* const endOfStreamReached ) const = 0;
     };
